port-probe: always cast SEND_DELAY value to guint64
See similar bug in GStreamer:
https://bugzilla.gnome.org/show_bug.cgi?id=740191

(cherry picked from commit a29610f25da4ce4843be2df87fb644be87e6d35a)

+++ src/mm-port-probe.c
@@ -1182,7 +1182,7 @@ serial_open_at (MMPortProbe *self)
 
         g_object_set (task->serial,
                       MM_PORT_SERIAL_SPEW_CONTROL,   TRUE,
-                      MM_PORT_SERIAL_SEND_DELAY,     (subsys == MM_PORT_SUBSYS_TTY ? task->at_send_delay : 0),
+                      MM_PORT_SERIAL_SEND_DELAY,     (guint64)(subsys == MM_PORT_SUBSYS_TTY ? task->at_send_delay : 0),
                       MM_PORT_SERIAL_AT_REMOVE_ECHO, task->at_remove_echo,
                       MM_PORT_SERIAL_AT_SEND_LF,     task->at_send_lf,
                       NULL);
